requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
beautifulsoup4==4.9.1
Django==3.0.6
django-bootstrap4==2.0.1
django-filter==2.2.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.4
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
